+++ frontend/src/components/questiontypes/videoupload_questiontype.tsx
@@ -4,20 +4,113 @@ import React, { useEffect, useState } from "react";
 
 import {
   deleteVideoUploadAnswer,
-  saveVideoUploadAnswer,
-  fetchVideoUploadAnswer,
 } from "@/actions/answers/videoUpload";
 import Logger from "@/logger/logger";
 import { UpdateAnswer } from "@/store/slices/answerSlice";
 import { useAppDispatch, useAppSelector } from "@/store/store";
-import { downloadFile } from "@/utils/helpers";
+import { downloadFile, storageSaveName } from "@/utils/helpers";
 
 import QuestionTypes, { DefaultQuestionTypeProps } from "./questiontypes";
 import { AwaitingChild } from "../layout/awaiting";
 import { SubmitButton } from "../submitButton";
+import { getSupabaseBrowserClient } from "@/supabase-utils/browserClient";
+import { saveAnswerN } from "@/actions/answers/answers";
+
+export interface VideoAnswerResponse {
+  answerid: string;
+  videoname: string;
+}
 
 export interface VideoUploadQuestionTypeProps extends DefaultQuestionTypeProps {
   maxfilesizeinmb: number;
+}
+
+export async function fetchVideoUploadAnswer(questionid: string) {
+  const supabase = getSupabaseBrowserClient();
+  const { data: userData, error: userError } = await supabase.auth.getUser();
+  if (userError) {
+    log.error(JSON.stringify(userError));
+  }
+  const user_id = userData.user!.id;
+
+  const { data: videoUploadData, error: videoUploadError } = await supabase
+    .rpc("fetch_video_upload_answer_table", {
+      question_id: questionid,
+      user_id: user_id,
+    })
+    .single<VideoAnswerResponse>();
+
+  if (videoUploadError) {
+    if (videoUploadError.code == "PGRST116") {
+      return null;
+    }
+    log.error(JSON.stringify(videoUploadError));
+    return null;
+  }
+  return { ...videoUploadData, userid: user_id };
+}
+
+export async function saveVideoUploadAnswer(
+  questionid: string,
+  formData: FormData,
+) {
+  const file = formData.get(questionid) as File;
+  const uploadFile = new File([file], storageSaveName(file.name), {
+    type: file.type,
+    lastModified: file.lastModified,
+  });
+  const bucket_name = `video-${questionid}`;
+  if (uploadFile) {
+    const { answerid, reqtype } = await saveAnswerN(questionid);
+    const supabase = getSupabaseBrowserClient();
+    if (reqtype == "created") {
+      const { error: insertAnswerError } = await supabase
+        .from("video_upload_answer_table")
+        .insert({
+          answerid: answerid,
+          videoname: uploadFile.name,
+        });
+      if (insertAnswerError) {
+        log.error("insertAnswerError " + JSON.stringify(insertAnswerError));
+      }
+      const { error: bucketError } = await supabase.storage
+        .from(bucket_name)
+        .upload(
+          `${(await supabase.auth.getUser()).data.user!.id}_${uploadFile.name}`,
+          uploadFile,
+        );
+      if (bucketError) {
+        log.error("bucketError: " + JSON.stringify(bucketError));
+      }
+    } else if (reqtype == "updated") {
+      const { data: oldVideoData, error: oldVideoError } = await supabase
+        .from("video_upload_answer_table")
+        .select("videoname")
+        .eq("answerid", answerid)
+        .single();
+      if (oldVideoError) {
+        log.error("oldVideoError " + JSON.stringify(oldVideoError));
+      }
+      const { error: updatedVideoError } = await supabase
+        .from("video_upload_answer_table")
+        .update({ videoname: uploadFile.name })
+        .eq("answerid", answerid);
+      if (updatedVideoError) {
+        log.error("updatedVideoError " + JSON.stringify(updatedVideoError));
+      }
+      const { error: updatedBucketError } = await supabase.storage
+        .from(bucket_name)
+        .update(
+          `${
+            (await supabase.auth.getUser()).data.user!.id
+          }_${oldVideoData?.videoname}`,
+          uploadFile,
+        );
+      if (updatedBucketError) {
+        log.error("updatedBucketError " + JSON.stringify(updatedBucketError));
+      }
+    }
+  }
 }
 
 const log = new Logger("VideoUploadQuestionType");
